test_project/requirements.txt
asgiref==3.6.0
Django==3.2.18
djangorestframework==3.14.0
et-xmlfile==1.1.0
gunicorn==20.1.0
joblib==1.2.0
numpy==1.21.6
openpyxl==3.1.2
pandas==1.3.5
python-dateutil==2.8.2
pytz==2022.7.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sqlparse==0.4.3
threadpoolctl==3.1.0
typing-extensions==4.5.0
